refactor: extract get_class_hierarchy cluster to QueryEngine (Phase 5)
Move get_class_hierarchy and 5 helpers from CppAnalyzer to QueryEngine:
- _resolve_base_key
- _lookup_class_infos
- _collect_hierarchy_node_data
- _should_skip_hierarchy_node
- _bfs_traverse_hierarchy
- get_class_hierarchy

CppAnalyzer keeps 1-line delegation for backward compatibility.
All state access in QueryEngine goes through self.analyzer back-reference.

All 1435 tests pass.
